+++ source/exe/main_common.cc
@@ -141,15 +141,13 @@ MainCommon::MainCommon(int argc, const char* const* argv)
       base_(options_, real_time_system_, default_test_hooks_, prod_component_factory_,
             std::make_unique<Runtime::RandomGeneratorImpl>(), platform_impl_.threadFactory()) {}
 
-std::string MainCommon::hotRestartVersion(uint64_t max_num_stats, uint64_t max_stat_name_len,
-                                          bool hot_restart_enabled) {
+std::string MainCommon::hotRestartVersion(uint64_t max_stat_name_len, bool hot_restart_enabled) {
 #ifdef ENVOY_HOT_RESTART
   if (hot_restart_enabled) {
-    return Server::HotRestartImpl::hotRestartVersion(max_num_stats, max_stat_name_len);
+    return Server::HotRestartImpl::hotRestartVersion(max_stat_name_len);
   }
 #else
   UNREFERENCED_PARAMETER(hot_restart_enabled);
-  UNREFERENCED_PARAMETER(max_num_stats);
   UNREFERENCED_PARAMETER(max_stat_name_len);
 #endif
   return "disabled";