+++ source/exe/main_common.h
@@ -99,8 +99,7 @@ public:
     base_.adminRequest(path_and_query, method, handler);
   }
 
-  static std::string hotRestartVersion(uint64_t max_num_stats, uint64_t max_stat_name_len,
-                                       bool hot_restart_enabled);
+  static std::string hotRestartVersion(uint64_t max_stat_name_len, bool hot_restart_enabled);
 
 private:
 #ifdef ENVOY_HANDLE_SIGNALS