+++ test/integration/server.cc
@@ -36,7 +36,6 @@ OptionsImpl createTestOptionsImpl(const std::string& config_path, const std::str
   test_options.setFileFlushIntervalMsec(std::chrono::milliseconds(50));
   test_options.setDrainTime(std::chrono::seconds(1));
   test_options.setParentShutdownTime(std::chrono::seconds(2));
-  test_options.setMaxStats(16384u);
 
   return test_options;
 }