+++ test/server/hot_restart_impl_test.cc
@@ -59,7 +59,7 @@ TEST_F(HotRestartImplTest, versionString) {
   // Tests that the version-string will be consistent and HOT_RESTART_VERSION,
   // between multiple instantiations.
   std::string version;
-  uint64_t max_stats, max_obj_name_length;
+  uint64_t max_obj_name_length;
 
   // The mocking infrastructure requires a test setup and teardown every time we
   // want to re-instantiate HotRestartImpl.
@@ -67,7 +67,6 @@ TEST_F(HotRestartImplTest, versionString) {
     setup();
     version = hot_restart_->version();
     EXPECT_TRUE(absl::StartsWith(version, fmt::format("{}.", HOT_RESTART_VERSION))) << version;
-    max_stats = options_.maxStats(); // Save this so we can double it below.
     max_obj_name_length = options_.statsOptions().maxObjNameLength();
     TearDown();
   }
@@ -75,13 +74,6 @@ TEST_F(HotRestartImplTest, versionString) {
   {
     setup();
     EXPECT_EQ(version, hot_restart_->version()) << "Version string deterministic from options";
-    TearDown();
-  }
-
-  {
-    ON_CALL(options_, maxStats()).WillByDefault(Return(2 * max_stats));
-    setup();
-    EXPECT_NE(version, hot_restart_->version()) << "Version changes when max-stats change";
     TearDown();
   }